+++ app/dao/sql/tables/UserToUserRelationTable.scala
@@ -6,9 +6,9 @@ import implicits._
 
 private[sql] class UserToUserRelationTable(tag: Tag) extends Table[UserToUserRelation](tag, "user_to_user_relation") {
 
-   def userFrom = column[Long]("user_from")
+   def userFrom = column[Int]("user_from")
 
-   def userTo = column[Long]("user_to")
+   def userTo = column[Int]("user_to")
 
    def relation = column[UsersRelationType]("relation_type")
