+++ app/models/persistient/Event.scala
@@ -2,7 +2,7 @@ package models.persistient
 
 final case class Event(
                    id: Long = 0L,
-                   ownerId: Long = 0L,
+                   ownerId: Int = 0,
                    dateMills: Long = System.currentTimeMillis(),
                    creationDateMills: Long = System.currentTimeMillis(),
                    maxMembers: Long = 0L,
