+++ app/models/persistient/EventToUser.scala
@@ -3,7 +3,7 @@ package models.persistient
 case class EventToUser(
 
                          eventId: Long = 0L,
-                         userId: Long = 0L,
+                         userId: Int = 0,
                          isNewForMember: Boolean = true
 
                       )
