+++ app/services/JWTCoderImpl.scala
@@ -35,7 +35,7 @@ class JWTCoderImpl @Inject()()(implicit val configuration: Configuration) extend
 
       val (username,secret,emailAddress) = getBasicCredentials(session)
 
-      val userId = session.getAs[Long](TokenRep.USER_ID).getOrElse(
+      val userId = session.getAs[Int](TokenRep.USER_ID).getOrElse(
          throw new RuntimeException("Can not get userId from private token")
       )
